+++ tests/configs/config_ERP_CORE.py
@@ -50,6 +50,11 @@ h_freq = None
 
 decode = True
 
+find_breaks = True
+min_break_duration = 10
+t_break_annot_start_after_previous_event = 3.0
+t_break_annot_stop_before_next_event = 1.5
+
 ica_reject = dict(eeg=350e-6, eog=500e-6)
 reject = dict(eeg=150e-6)